fix(tfmigrate-plan): delete old plan file in GCS when terraform plan fails

+++ tfmigrate-plan/main.sh
@@ -4,7 +4,12 @@ set -eu
 set -o pipefail
 
 echo "===> Delete old plan file to prevent the accident" >&2
-aws s3 rm "s3://$S3_BUCKET_NAME_PLAN_FILE/$PR_NUMBER/$TFACTION_TARGET/tfplan.binary" || :
+if [ -n "${S3_BUCKET_NAME_PLAN_FILE:-}" ]; then
+	aws s3 rm "s3://$S3_BUCKET_NAME_PLAN_FILE/$PR_NUMBER/$TFACTION_TARGET/tfplan.binary" || :
+fi
+if [ -n "${GCS_BUCKET_NAME_PLAN_FILE:-}" ]; then
+	gsutil rm "gs://$GCS_BUCKET_NAME_PLAN_FILE/$PR_NUMBER/$TFACTION_TARGET/tfplan.binary" || :
+fi
 
 if [ ! -f .tfmigrate.hcl ]; then
 	sed "s|%%TARGET%%|$TFACTION_TARGET|g" \
